- add Registry module to keep track of the number to NFT relationship - add Collection module to keep track of the collection supply and avoid having a circular dependency between the Registry module and the Water Cooler module - update Water Cooler Moduler to add NFTs to registry when they are created - update tests to support the changes that have been made

+++ sources/mint.move
@@ -232,7 +232,7 @@ module galliun::mint {
         };
         nfts.destroy_empty();
 
-        if (warehouse.nfts.length() == water_cooler.supply()) {
+        if (warehouse.nfts.length() as u64 == water_cooler.supply()) {
             warehouse.is_initialized = true;
         };
     }
